mediamanager/requirements.txt
aioredis==1.0.0
argon2-cffi==20.1.0
asgiref==3.2.10
astroid==2.4.1
async-generator==1.10
async-timeout==2.0.1
attrs==17.4.0
autobahn==18.3.1
Automat==0.6.0
autopep8==1.5.4
black==21.5b0
bleach==3.3.0
boto3==1.9.96
botocore==1.12.96
certifi==2020.6.20
cffi==1.14.3
channels==2.0.2
channels-redis==2.1.1
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.1.1
daphne==2.1.0
decorator==4.4.2
defusedxml==0.7.0rc1
diff-match-patch==20200713
dj-database-url==0.5.0
Django==3.1.1
django-braces==1.14.0
django-cors-headers==3.5.0
django-crispy-forms==1.7.2
django-debug-toolbar==2.1
django-excel-response2==2.0.8
django-filter==2.4.0
django-heroku==0.3.1
django-import-export==2.5.0
django-oauth-toolkit==1.3.2
django-rest-knox==4.1.0
django-s3direct==1.1.8
django-six==1.0.4
django-storages==1.7.1
django-webpack-loader==0.6.0
djangorestframework==3.11.1
djangorestframework-simplejwt==4.6.0
docutils==0.14
dodgy==0.2.1
drf-social-oauth2==1.0.8
entrypoints==0.3
et-xmlfile==1.1.0
flake8==3.8.4
flake8-polyfill==1.0.2
gunicorn==19.9.0
hiredis==2.0.0
hyperlink==18.0.0
ibm-cloud-sdk-core==3.5.2
ibm-watson==5.1.0
idna==2.10
imageio==2.9.0
imageio-ffmpeg==0.4.4
incremental==17.5.0
ipykernel==5.5.0
ipython==7.20.0
ipywidgets==7.6.3
isort==4.3.21
itypes==1.2.0
Jinja2==2.11.2
jmespath==0.9.3
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.11
jupyter-console==6.2.0
jupyter-core==4.7.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.0
lazy-object-proxy==1.4.3
MarkupPy==1.14
mccabe==0.6.1
mistune==0.8.4
moviepy==1.0.3
msgpack==0.5.6
mypy-extensions==0.4.3
nbclient==0.5.2
nbconvert==6.0.7
nbformat==5.1.2
nest-asyncio==1.5.1
notebook==6.2.0
numpy==1.20.1
oauthlib==3.1.0
odfpy==1.4.1
openpyxl==3.0.7
pandocfilters==1.4.3
pathspec==0.8.1
pep8-naming==0.10.0
Pillow==5.2.0
proglog==0.1.9
prometheus-client==0.9.0
prospector==1.3.1
psycopg2==2.8.6
psycopg2-binary==2.8.6
pycodestyle==2.6.0
pycosat==0.6.3
pydocstyle==6.0.0
pyflakes==2.2.0
PyJWT==1.7.1
pylint==2.5.3
pylint-celery==0.3
pylint-django==2.1.0
pylint-flask==0.6
pylint-plugin-utils==0.6
pyrsistent==0.17.3
python-dateutil==2.8.0
python3-openid==3.2.0
pytz==2020.1
pyzmq==22.0.3
qtconsole==5.0.2
QtPy==1.9.0
regex==2021.4.4
requests==2.24.0
requests-oauthlib==1.3.0
requirements-detector==0.7
ruamel.yaml==0.16.12
s3transfer==0.2.0
scipy==1.6.3
screen==1.0.1
Send2Trash==1.5.0
setoptconf==0.2.0
six==1.15.0
snowballstemmer==2.1.0
social-auth-app-django==4.0.0
social-auth-core==3.3.3
sqlparse==0.3.1
tablib==3.0.0
terminado==0.9.2
testpath==0.4.4
toml==0.10.1
tornado==6.1
tqdm==4.60.0
Twisted==17.9.0
txaio==2.9.0
urllib3==1.25.10
websocket-client==0.48.0
whitenoise==4.1.2
widgetsnbextension==3.5.1
wrapt==1.12.1
xlrd==2.0.1
xlwt==1.3.0
yapf==0.31.0
youtube-dl==2021.2.10
youtube-transcript-api==0.4.1
